requirements.txt
Flask
Flask-Cors
Flask-SQLAlchemy
psycopg2-binary
python-dotenv
gunicorn